Consolidate entry points
